add xor controller and aggregator part
