+++ nvflare/edge/examples/jobs/hello_mobile/meta.json
