+++ nvflare/edge/examples/jobs/xor_mobile/app/config/config_fed_client.json
