+++ nvflare/edge/examples/jobs/xor_mobile/app/custom/model.py
