+++ nvflare/edge/examples/jobs/xor_mobile/meta.json
