+++ nvflare/edge/web/models/job_response.py
@@ -17,7 +17,6 @@ from nvflare.edge.web.models.base_model import BaseModel
 
 
 class JobResponse(BaseModel):
-
     def __init__(
         self,
         status: str,
